para_athlete_project/requirements.txt
pandas
numpy
scikit-learn
joblib
fastapi
uvicorn
requests
google-genai
python-dotenv
